From Bjoern:
I would like to request the prefix MRO. My first prefix request ever, yeay!

- Title: MHC restriction ontology
- Short description: An application ontology capturing how MHC restriction is defined in experiments, spanning exact protein complexes, individual protein chains, serotypes, haplotypes and mutant molecules, as well as evidence for MHC restrictions
- Download (for now): https://github.com/IEDB/MRO; (we will rely on PURLs once domain approved)
- Homepage is the same github page
- Contact is me, Bjoern Peters, bpeters@liai.org
- Tracker is on the github page

I do want to point out that we are actively working with several other people in this domain, specifically PRO, and we have also been in contact with not currently maintained MHC ontology. We have also been in contact with Marcus for the evidence ontology. The ontology is both practical useful for our work in the IEDB right now, and could serve as a broader template how to deal with different expeirmental evidence to identify members of diverse protein families.