Update fp-001-open.md
##Principle Name 
...

I unnecessary as the name is give in the top level of page. I'm doing this as PR to have someone else review that my change won't break anything else. If it is accepted I will directed edit the rest of the principle pages in the same way. Longer discussion at https://github.com/OBOFoundry/OBOFoundry.github.io/issues/302