Add optional separate aberowl_id for an ontology.
NOTES: in the case of the Relations Ontology, the OBO id is 'RO' but the AberOWL
id is 'RELO'. Using the OBO id of 'RO' to link to the AberOWL page, sends the
user to the Radiomics Ontology page of AberOWL. This change allows a contributor
to request a particular AberOWL id/redirect link.

Resolves #1053