Create cido.md
The CIDO ontology has been posted to obo-discuss email list for discussion for over 2 weeks and has been agreed from the discussion that this is a timely ontology to study coronavirus diseases including COVID-19. A note of change is that we transferred the CIDO github organization location from github.com/biomedontology to the newly generated https://github.com/CIDO-ontology organization. Such a transfer is to better support its organization with more options (such as generating new repositories).