New "about" page
This is mostly taken from the "wall of text" on the current home page. The plan is to simplify the text on the home page and move most of the details to this new About page (see #241).
I think it could still use some polishing, but this is a starting point.
Note that at the moment this page is not linked anywhere. Will do that later.