Update cdno.md
Update the scope and changed the name of the ontology from Crop Dietary Nutrition Ontology to Compositional Dietary Nutrition Ontology. The changes in the name included the modification of the design patterns, changing 'plant structure' for 'material entity' in the 'concentration of dietary nutritional component' class. This changes were discussed in the issue "Process to change Crop for Compositional in CDNO #1483" (https://github.com/OBOFoundry/OBOFoundry.github.io/issues/1483).