Assign Sue as contact for MP due to FP-11
FP11 is the locus of authority rule, which means that the contact for an ontology can't be a mailing list.  @sbello is listed as the person who's made the most contributions (and her email is public on github) so this seems reasonable. We can change it to someone else if that's more appropriate, instead.