Standardize YAML in ontology metadata frontmatter
This PR does the following:

1. Introduces a Python script util/standardize_metadata.py that sorts and properly indents all of the metadata in all ontology metadata files' frontmatter
2. Adds a unit test to ensure that this metadata is sorted

Why do this? It makes it much easier to programatically edit these files without creating massive diffs (after the massive diff in this PR, at least). There are several ways we might want to automatically maintain these metadata, such as the list of dependencies for each ontology.