Create permissible-metadata-content (#1988)
* Create permissible-metadata-content

NOTE: This file begins the process of collecting information needed to fill out YAML files containing ontology metadata. The JSON schema is used for this purpose because it (presumably) contains all the fields that can be filled in. It will be prettied up as we go along.

* Create permissible-metadata-content.md

* Information for the 'publications' field

Note: in the (near?) future we will allow CURIEs in addition to URIs. When that happens, the initial sentence under 'id' will be revised to indicate such.

* Removed some extraneous info

Attempting to focus on fields that the user can change. This file is intended to guide the construction of an FAQ on metadata content.

* Update permissible-metadata-content.md

Corrections and clarifications.

* Update permissible-metadata-content.md

Better formatting.

* Delete permissible-metadata-content

Not needed