Home page table cleanup (#2041)
* Trim whitespace from description

When the last character of the description is a space, it previously looked very weird with the ellipses

* Add scope to table headers

* Improve quick access

- Group with a button group
- Add role annotations

* Make sort buttons smaller

* Add disabled button when missing tracker

* Update GitHub cell

Improve handling for when not a github repository. Previously it made too many assumptions

* Improve contact button

This adds a unclickable button when no contact is available, rather than creating an undefined element

* Improve publication button

This makes a greyed out button if no publication is displayed

* Improve handling desription

Previously the code assumed descriptions were always available. This isn't the case for some ontologies, like BILA.

* Improve handling of license

Previously, it did some nonsense with alt labels for the licenses with no URL. I'm still not happy with using hidden text for sorting, but this is maintained for now.

* Remove empty class and fix spacing

* Clean up layouts

- remove unnecessary div nesting
- lint

* Better handle inactive ontologies

- add missing logic for orphaned ontologies
- move text over so ontology prefix column has more space