Add missing products to several ontologies
I wrote a script that tries to fill these in by re-writing the file extensions then running a HEAD request with redirects to see if the file exists. I then went through and **manually checked** that actual files exist for all of the changes proposed here.

Rationale: I want JSON exports of everything, and some were available and I was wasting tons of time converting OWL/OBO with ROBOT - much better to have it annotated directly so these can be automatically consumed.