Add new optional metadata field for pull request added (#2146)
* Add new metadata field for pull request added

This PR adds the ability to encode the pull request when an ontology was added to the OBO Foundry GitHub repository.

It's true that this repository does not well-reflect ontologies that were pulled over from SVN, and therefore the oldest ontologies will not have good entries for this. Please do not let this deter the acceptance of this PR.

This PR will allow an alternate venue of identifying the date added for new repositories. It can be made a requirement for new ontologies later. It can also be a nice curation task as a "good first issue"

* Add NOR field

* Update t4fs.md

* Update ontology_detail.html

* Update ontology_detail.html

Co-authored-by: Nico Matentzoglu <nicolas.matentzoglu@gmail.com>