Remove unused IDs
`id` is not a privileged or especially meaningful key for jekyll. These do not do anything and are confusing, so best to remove them.

That being said, some pages, such as the OBO principles, do use the content in the `id` key for display so they should not be removed.