Update fp-009-users.md to emphasize requirements for links
Revised the document to draw attention to the need for links in usages declarations (this requirement was already present, but was spelled out in the examples). The examples as they were prior to revision were general and contained important information, so these were moved to the R & R section, while the Examples section was revised to give concrete examples.