Change column header (#2618)
The second-to-last column of the ontology table is currently labeled "Re-Use", but it contains both the dashboard results and the license. I was in favor of separating those into separate columns, but there was concern that that would make the table too wide. The proposed compromise was to change the column header, as I have done.
address #2611