reorder fields in a more logical way
why did these records start with affiliation, with name buried somewhere in the middle? that makes it much tricker to correctly make edits (adding or removing someone).

i reordered the fields (with the help of a Claude-generated script) to put name first and affiliation list, in hopes of making this file more maintainable both in the short term (i need to move some members to alumni) and longer term.

lmk if anything looks wonky; i only spot-checked.