update Slack link (#2788)
* update Slack link

old link had expired. created a new supposedly never-expiring link.

* Update Slack link

(the Slack link on this home page was the generic one that only works if you're already a member - was there a reason we decided not to include the invite link here? if so, please remind me!)

* update Slack link

also correct which steps are internal to the EWG

* Update Slack link